fixate caps to the nearest desired width, height & framerate
Harden the input selection process a bit. The width, height or framerate fields
in a Gst cap could possibly contain non fixated values. For example, framerate
could be a list of fractions (issue #3).

fix: make --input-{width, height, framerate} required with sane defaults
and try to fixate to the neareast of those values.
Now, --input-{width, height, framerate} might not exactly match the chosen
inputs, but at least we'll choose something close to it. Which is what we
usually want anyways.